for those who

+++ kernel/icxxabi.cpp
@@ -1,5 +1,7 @@
+// Cribbed from osdev.org
 #include <kernel/icxxabi.h>
 #include <kernel/util/Spinlock.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -13,12 +15,12 @@ void *__dso_handle = 0; // Attention! Optimally, you should remove the '= 0'
 
 Spinlock cxa_spinlock;
 
-int __cxa_guard_acquire(long long int *) {
+int __cxa_guard_acquire(int64_t *) {
 	// cxa_spinlock.acquire();
 	return 1;
 }
 
-void __cxa_guard_release(long long int *) { /*cxa_spinlock.release();*/
+void __cxa_guard_release(int64_t *) { /*cxa_spinlock.release();*/
 }
 
 void __cxa_pure_virtual() {
